backend/requirements.txt
python-dotenv
fastapi
openai
uvicorn[standard]
authlib
pydantic
pydantic-settings
itsdangerous
httpx